Fix runtime issues associated with lib split
