+++ SpectabisNext/Controls/GameTileView/GameTileView.xaml
