+++ SpectabisNext/Controls/PageIcon/PageIcon.xaml
