+++ SpectabisNext/Controls/PageIcon/PageIcon.xaml.cs
