+++ SpectabisNext/Repositories/PageRepository.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Linq;
-using SpectabisNext.Controls;
 using SpectabisNext.Views;
 using SpectabisUI.Controls;
