+++ SpectabisNext/Views/FirstTimeWizard.xaml.cs
@@ -1,4 +1,3 @@
-using SpectabisNext.Controls;
 using SpectabisNext.Services;
 using SpectabisUI.Controls;
