+++ SpectabisNext/Views/Settings.xaml.cs
@@ -1,7 +1,6 @@
 using System;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
-using SpectabisNext.Controls;
 using SpectabisUI.Controls;
 
 namespace SpectabisNext.Views
